extract comment handlers
